Replace Profession.java with Occupation.java
