+++ src/main/java/seedu/address/model/person/Occupation.java
@@ -1,0 +1,59 @@
+package seedu.address.model.person;
+
+import static java.util.Objects.requireNonNull;
+import static seedu.address.commons.util.AppUtil.checkArgument;
+
+/**
+ * Represents a Person's  in the address book.
+ * Guarantees: immutable; is valid as declared in {@link #isValidOccupation(String)}
+ */
+public class Occupation {
+
+    public static final String MESSAGE_CONSTRAINTS =
+            "Occupation should only contain alphanumeric characters and spaces, and it should not be blank";
+
+    /*
+     * The first character of the  must not be a whitespace,
+     * otherwise " " (a blank string) becomes a valid input.
+     */
+    public static final String VALIDATION_REGEX = "[\\p{Alnum}][\\p{Alnum} ]*";
+
+    public final String occupation;
+    public String value;
+
+    /**
+     * Constructs a {@code Occupation}.
+     *
+     * @param occupation A valid occupation.
+     */
+    public Occupation(String occupation) {
+        requireNonNull(occupation);
+        checkArgument(isValidOccupation(occupation), MESSAGE_CONSTRAINTS);
+        this.occupation = occupation;
+    }
+
+    /**
+     * Returns true if a given string is a valid occupation.
+     */
+    public static boolean isValidOccupation(String test) {
+        return test.matches(VALIDATION_REGEX);
+    }
+
+    @Override
+    public String toString() {
+        return occupation;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof Occupation // instanceof handles nulls
+                && occupation.equals(((Occupation) other).occupation)); // state check
+    }
+
+    @Override
+    public int hashCode() {
+        return occupation.hashCode();
+    }
+
+}
